Fix test code: check ./set-of-emails/obsoleted directory

+++ spec/obsoleted/sisimai_spec.rb
@@ -32,6 +32,7 @@ describe Sisimai do
   describe '.make' do
     context 'valid JSON file' do
       samplejsons.each do |e|
+        break unless Dir.exist?('./set-of-emails/obsoleted')
         jf = File.open('./set-of-emails/obsoleted/' + e, 'r')
         js = jf.read
         jf.close
